requirements.txt
openai==0.27.8
peewee==3.17.0
selenium==4.15.2
newspaper3k==0.2.8
readability-lxml==0.8.1
langdetect==1.0.9
retry==0.9.2
prompt-toolkit==3.0.40
PyYAML==6.0.1
rich==13.7.0
pypdf2==3.0.1
